+++ bayesflow/experimental/self_consistency/__init__.py
@@ -1,0 +1,4 @@
+from .semi_supervised_approximator import SemiSupervisedApproximator
+from .self_consistency_loss import SelfConsistencyLoss
+
+__all__ = ["SemiSupervisedApproximator", "SelfConsistencyLoss"]
