+++ bayesflow/experimental/self_consistency/datasets/__init__.py
@@ -1,0 +1,4 @@
+from .multi_dataset import MultiDataset
+from .self_consistency_dataset import SelfConsistencyDataset
+
+__all__ = ["MultiDataset", "SelfConsistencyDataset"]
