+++ bayesflow/experimental/self_consistency/datasets/multi_dataset.py
