--- bayesflow/schedules/__init__.py
@@ -1,4 +1,0 @@
-from .weight_schedule import WeightSchedule, ConstantSchedule, LinearSchedule
-from ..utils._docs import _add_imports_to_all
-
-_add_imports_to_all(include_modules=[])
